Address review: export+rename brand key, drop errors/index.ts barrel
- Rename the brand key to 'instanceof conjure-client#ConjureError' and export it
  so it has a single source of truth (the test imports it rather than hardcoding).
- Remove the src/errors/index.ts 'export *' barrel; import ConjureError et al.
  directly from ./errors/error at the four call sites (src/index.ts, types/result.ts,
  fetchBridge.ts, fetchBridgeTests.ts).
